Provide RepositoryBranch metadata for package
As documented in nuget (https://learn.microsoft.com/en-us/nuget/reference/msbuild-targets) you can provide
not only the repo URL, type and commit, but also the branch. But this information is not populated by default
by the SourceLink package, neither the ReproducibleBuilds (unless the PR at https://github.com/dotnet/reproducible-builds/pull/27 is merged).

So just bring the code in that PR here until that's merged and shipped.

Fixes #23